+++ isolde/src/restraints_cpp/distance_restraints.cpp
@@ -76,28 +76,28 @@ double Distance_Restraint::radius() const
         * (LINEAR_RESTRAINT_MAX_RADIUS-LINEAR_RESTRAINT_MIN_RADIUS) + LINEAR_RESTRAINT_MIN_RADIUS;
 }
 
-void Distance_Restraint::target_transform(double *rot44) const
+void Distance_Restraint::target_transform(float *rot44) const
 {
-    double scale = get_target() / distance();
+    float scale = get_target() / distance();
     _bond_transform(rot44, radius(), scale);
 }
 
-void Distance_Restraint::bond_cylinder_transform(double *rot44) const
+void Distance_Restraint::bond_cylinder_transform(float *rot44) const
 {
     _bond_transform(rot44, 1.0, 1.0);
 }
 
-void Distance_Restraint::_bond_transform(double *rot44, double radius, double length_scale) const
+void Distance_Restraint::_bond_transform(float *rot44, float radius, float length_scale) const
 {
     const Coord &c0 = atoms()[0]->coord();
     const Coord &c1 = atoms()[1]->coord();
-    double xyz0[3], xyz1[3];
+    float xyz0[3], xyz1[3];
     for (size_t i=0; i<3; ++i)
     {
         xyz0[i] = c0[i];
         xyz1[i] = c1[i];
     }
-    geometry::bond_cylinder_transform_gl<double>(xyz0, xyz1, radius, length_scale, rot44);
+    geometry::bond_cylinder_transform_gl<float>(xyz0, xyz1, radius, length_scale, rot44);
 }
